+++ Flow Graph Library/Fuzzing Flow Graphs/RDS_BPSK_HackRF_Fields.py
@@ -34,6 +34,7 @@ class RDS_BPSK_HackRF_Fields(gr.top_block):
         ##################################################
         self.transmit_interval = transmit_interval = 1
         self.stereo_gain = stereo_gain = .3
+        self.serial = serial = "0"
         self.samp_rate = samp_rate = 1e6
         self.rds_gain = rds_gain = .5
         self.pilot_gain = pilot_gain = .3
@@ -57,7 +58,7 @@ class RDS_BPSK_HackRF_Fields(gr.top_block):
         # Blocks
         ##################################################
         self.osmosdr_sink_0 = osmosdr.sink(
-            args="numchan=" + str(1) + " " + ''
+            args="numchan=" + str(1) + " " + "hackrf=" + str(serial)
         )
         self.osmosdr_sink_0.set_time_unknown_pps(osmosdr.time_spec_t())
         self.osmosdr_sink_0.set_sample_rate(1e6)
@@ -126,6 +127,12 @@ class RDS_BPSK_HackRF_Fields(gr.top_block):
 
     def set_stereo_gain(self, stereo_gain):
         self.stereo_gain = stereo_gain
+
+    def get_serial(self):
+        return self.serial
+
+    def set_serial(self, serial):
+        self.serial = serial
 
     def get_samp_rate(self):
         return self.samp_rate