+++ Flow Graph Library/Single-Stage Flow Graphs/RDS_BPSK_HackRF_File_Source.py
@@ -33,6 +33,7 @@ class RDS_BPSK_HackRF_File_Source(gr.top_block):
         # Variables
         ##################################################
         self.stereo_gain = stereo_gain = .3
+        self.serial = serial = "0"
         self.rds_gain = rds_gain = .5
         self.pilot_gain = pilot_gain = .3
         self.outbuffer = outbuffer = 10
@@ -47,7 +48,7 @@ class RDS_BPSK_HackRF_File_Source(gr.top_block):
         # Blocks
         ##################################################
         self.osmosdr_sink_0 = osmosdr.sink(
-            args="numchan=" + str(1) + " " + ''
+            args="numchan=" + str(1) + " " + "hackrf=" + str(serial)
         )
         self.osmosdr_sink_0.set_time_unknown_pps(osmosdr.time_spec_t())
         self.osmosdr_sink_0.set_sample_rate(1e6)
@@ -144,6 +145,12 @@ class RDS_BPSK_HackRF_File_Source(gr.top_block):
 
     def set_stereo_gain(self, stereo_gain):
         self.stereo_gain = stereo_gain
+
+    def get_serial(self):
+        return self.serial
+
+    def set_serial(self, serial):
+        self.serial = serial
 
     def get_rds_gain(self):
         return self.rds_gain